Move getFighters to /services & update import paths
