+++ scraping/services/getFighters.js
@@ -1,10 +1,10 @@
 import sleepWithLog from 'sleep-dots'
-import { logSuccess } from './utils/log.js'
-import { readDBFile } from './utils/db.js'
-import { REQUEST_DELAY } from '../octagon-api.config.js'
-import scrapeFighter from './scrapers/scrapeFighter.js'
-import timeFormatter from './utils/timeFormatter.js'
-import { RANKINGS_DB_NAME } from './constants/names.js'
+import { logSuccess } from '../utils/log.js'
+import { readDBFile } from '../utils/db.js'
+import { REQUEST_DELAY } from '../../octagon-api.config.js'
+import scrapeFighter from '../scrapers/scrapeFighter.js'
+import timeFormatter from '../utils/timeFormatter.js'
+import { RANKINGS_DB_NAME } from '../constants/names.js'
 
 const getFighters = async () => {
   const start = performance.now()
